Moved repository interface to different namespace
